+++ example/example.php
@@ -4,13 +4,16 @@
  * of the SEPA file format to learn more about using SEPA files.
  */
 
-require_once 'Sephpa.php';
+use AbcAeffchen\Sephpa\SephpaCreditTransfer;
+use AbcAeffchen\Sephpa\SephpaDirectDebit;
 
-// generate a SepaCreditTransfer object (pain.001.002.03). Here you can not use direct debit!
-$creditTransferFile = new Sephpa('Initiator Name', 'MessageID-1234', SEPA_PAIN_001_002_03);
+require_once '../src/Sephpa.php';
 
-// at least one in every SepaXmlFile (of type CT). No limit.
-$creditTransferCollection = $creditTransferFile->addCreditTransferCollection(array(
+// generate a SepaCreditTransfer object (pain.001.002.03).
+$creditTransferFile = new SephpaCreditTransfer('Initiator Name', 'MessageID-1234', SephpaCreditTransfer::SEPA_PAIN_001_002_03);
+
+// at least one in every SEPA file
+$creditTransferCollection = $creditTransferFile->addCollection(array(
                     // needed information about the payer
                         'pmtInfId'      => 'PaymentID-1234',    // ID of the payment collection
                         'dbtr'          => 'Name of Debtor2',   // (max 70 characters)
@@ -24,7 +27,7 @@ $creditTransferCollection = $creditTransferFile->addCreditTransferCollection(arr
                         'ultmtDebtr'    => 'Ultimate Debtor Name'   // just an information, this do not affect the payment (max 70 characters)
                     ));
                     
-// at least one in every CreditTransferCollection. No limit.
+// at least one in every CreditTransferCollection
 $creditTransferCollection->addPayment(array(
                     // needed information about the one who gets payed
                         'pmtId'     => 'TransferID-1234-1',     // ID of the payment (EndToEndId)
@@ -38,22 +41,14 @@ $creditTransferCollection->addPayment(array(
                         'rmtInf'    => 'Remittance Information' // unstructured information about the remittance (max 140 characters)
                     ));
 
+$creditTransferFile->storeSepaFile();
 
 
-// save the file on the server
-$ct = fopen('credit_transfer.xml', 'w');
-fwrite($ct, $creditTransferFile->generateXml());
-fclose($ct);
-// or download the file without saving it on the server
-header('Content-Disposition: attachment; filename="credit_transfer.xml"');
-print $creditTransferFile->generateXml();
+// generate a SepaDirectDebit object (pain.008.002.02).
+$directDebitFile = new SephpaDirectDebit('Initiator Name', 'MessageID-1235', SephpaDirectDebit::SEPA_PAIN_008_002_02);
 
-
-// generate a SepaDirectDebit object (pain.008.002.02). Here you can not use credit transfer!
-$directDebitFile = new Sephpa('Initiator Name', 'MessageID-1235', SEPA_PAIN_008_002_02);
-
-// at least one in every SepaXmlFile (of type DD). No limit.
-$directDebitCollection = $directDebitFile->addDirectDebitCollection(array(
+// at least one in every SEPA file. No limit.
+$directDebitCollection = $directDebitFile->addCollection(array(
                     // needed information about the payer
                         'pmtInfId'      => 'PaymentID-1235',        // ID of the paymentcollection
                         'lclInstrm'     => 'CORE',                  // only 'CORE' or 'B2B'
@@ -61,7 +56,7 @@ $directDebitCollection = $directDebitFile->addDirectDebitCollection(array(
                         'cdtr'          => 'Name of Creditor',      // (max 70 characters)
                         'iban'          => 'DE87200500001234567890',// IBAN of the Creditor
                         'bic'           => 'BELADEBEXXX',           // BIC of the Creditor
-                        'ci'            => 'DE00ZZZ00099999999',    // Creditor-Identifier
+                        'ci'            => 'DE98ZZZ09999999999',    // Creditor-Identifier
                     // optional
                         'ccy'           => 'EUR',                   // Currency. Default is 'EUR'
                         'btchBookg'     => 'true',                  // BatchBooking, only 'true' or 'false'
@@ -89,18 +84,10 @@ $directDebitCollection->addPayment(array(
                         // only use this if 'amdmntInd' is 'true'. at least one must be used
                         'orgnlMndtId'           => 'Original-Mandat-ID',
                         'orgnlCdtrSchmeId_nm'   => 'Creditor-Identifier Name',
-                        'orgnlCdtrSchmeId_id'   => 'Creditor-Identifier ID',
+                        'orgnlCdtrSchmeId_id'   => 'DE98AAA09999999999',
                         'orgnlDbtrAcct_iban'    => 'DE87200500001234567890',// Original Debtor Account
                         'orgnlDbtrAgt'          => 'SMNDA'          // only 'SMNDA' allowed if used
 ));
 
-// save the file on the server
-//$dd = fopen('direct_debit.xml', 'w');
-//fwrite($dd, $directDebitFile->generateXml());
-//fclose($dd);
-// or download the file without saving it on the server
-header('Content-Disposition: attachment; filename="direct_debit.xml"');
-print $directDebitFile->generateXml();
+$directDebitFile->downloadSepaFile();
 
-
-?>