+++ src/SephpaCreditTransfer.php
@@ -1,0 +1,88 @@
+<?php
+/**
+ * Sephpa
+ *  
+ * @license MIT License
+ * @copyright © 2014 Alexander Schickedanz
+ * @link      https://github.com/AbcAeffchen/Sephpa
+ *
+ * @author  Alexander Schickedanz <abcaeffchen@gmail.com>
+ */
+
+namespace AbcAeffchen\Sephpa;
+use AbcAeffchen\SepaUtilities\SepaUtilities;
+
+require_once 'Sephpa.php';
+
+/**
+ * Base class for both credit transfer and direct debit
+ */
+class SephpaCreditTransfer extends Sephpa
+{
+    const SEPA_PAIN_001_002_03 = SepaUtilities::SEPA_PAIN_001_002_03;
+    const SEPA_PAIN_001_003_03 = SepaUtilities::SEPA_PAIN_001_003_03;
+    /**
+     * @type string INITIAL_STRING_CT Initial sting for credit transfer payment-collections.001.002.03
+     */
+    const INITIAL_STRING_PAIN_001_002_03 = '<?xml version="1.0" encoding="UTF-8"?><Document xmlns="urn:iso:std:iso:20022:tech:xsd:payment-collections.001.002.03" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="urn:iso:std:iso:20022:tech:xsd:payment-collections.001.002.03 payment-collections.001.002.03.xsd"></Document>';
+    /**
+     * @type string INITIAL_STRING_CT Initial sting for credit transfer payment-collections.001.003.03
+     */
+    const INITIAL_STRING_PAIN_001_003_03 = '<?xml version="1.0" encoding="UTF-8"?><Document xmlns="urn:iso:std:iso:20022:tech:xsd:payment-collections.001.003.03" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="urn:iso:std:iso:20022:tech:xsd:payment-collections.001.003.03 payment-collections.001.003.03.xsd"></Document>';
+    /**
+     * Creates a SepaXmlFile object and sets the head data
+     *
+     * @param string $initgPty The name of the initiating party
+     * @param string $msgId    The unique id of the file
+     * @param int    $type     Sets the type and version of the sepa file. Use the SEPA_PAIN_* constants
+     * @param bool   $checkAndSanitize
+     * @throws SephpaInputException
+     */
+    public function __construct($initgPty, $msgId, $type, $checkAndSanitize = true)
+    {
+        parent::__construct($initgPty, $msgId, $type, $checkAndSanitize);
+
+        $this->xmlType = 'CstmrCdtTrfInitn';
+
+        switch($type)
+        {
+            case self::SEPA_PAIN_001_002_03:
+                $this->xml = simplexml_load_string(self::INITIAL_STRING_PAIN_001_002_03);
+                $this->type = self::SEPA_PAIN_001_002_03;
+                break;
+            case self::SEPA_PAIN_001_003_03:
+                $this->xml = simplexml_load_string(self::INITIAL_STRING_PAIN_001_003_03);
+                $this->type = self::SEPA_PAIN_001_002_03;
+                break;
+            default:
+                throw new SephpaInputException('You choose an invalid SEPA file version. Please use the SEPA_PAIN_001_* constants.');
+        }
+    }
+
+    /**
+     * Adds a new collection of credit transfers and sets main data
+     *
+     * @param mixed[] $transferInfo Required keys: 'pmtInfId', 'dbtr', 'iban', ('bic' only payment-collections.001.002.03);
+     *                              optional keys: 'ccy', 'btchBookg', 'ctgyPurp', 'reqdExctnDt', 'ultmtDbtr'
+     * @throws SephpaInputException
+     * @return SepaPaymentCollection
+     */
+    public function addCollection(array $transferInfo)
+    {
+        switch($this->type)
+        {
+            case self::SEPA_PAIN_001_002_03:
+                $paymentCollection = new SepaCreditTransfer00100203($transferInfo, $this->checkAndSanitize, $this->sanitizeFlags);
+                break;
+            case self::SEPA_PAIN_001_003_03:
+                $paymentCollection = new SepaCreditTransfer00100303($transferInfo, $this->checkAndSanitize, $this->sanitizeFlags);
+                break;
+            default:
+                throw new SephpaInputException('You choose an invalid SEPA file version. Please use the SEPA_PAIN_001_* constants.');
+        }
+        $this->paymentCollections[] = $paymentCollection;
+        
+        return $paymentCollection;
+    }
+
+}